requirements.txt
pydantic>=2.0.0
numpy>=1.24.0
matplotlib>=3.6.0
openai>=1.0.0
python-dotenv>=1.0.0